Criando um menu iterativo usando os conceitos de classes aprendidos até o momento.
