+++ introducao/fundamentos_csharp/src/ProgramacaoOrientadaObjeto/1_Namespaces/AssuntoNamespace.cs
@@ -1,0 +1,10 @@
+namespace ProgramacaoOrientadaObjeto.Namespaces;
+
+public class AssuntoNamespace(string titulo) : Assunto(titulo)
+{
+    public override void ExecutarDemo()
+    {
+        Console.WriteLine(Pessoa.Descricao());
+        Console.WriteLine(Entidades.Pessoa.Descricao());
+    }
+}
