+++ introducao/fundamentos_csharp/src/ProgramacaoOrientadaObjeto/2_Classes/ModeloClasse.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace ProgramacaoOrientadaObjeto.Classes;
 
 public class ModeloClasse
