+++ introducao/fundamentos_csharp/src/ProgramacaoOrientadaObjeto/2_Classes/ModeloStatic.cs
@@ -1,6 +1,4 @@
-using System;
-
-namespace ProgramacaoOrientadaObjeto._2_Classes;
+namespace ProgramacaoOrientadaObjeto.Classes;
 
 public static class ModeloStatic
 {
