+++ introducao/fundamentos_csharp/src/ProgramacaoOrientadaObjeto/Assunto.cs
@@ -1,0 +1,16 @@
+namespace ProgramacaoOrientadaObjeto;
+
+public abstract class Assunto
+{
+    public string Titulo { get; set; }
+
+    protected Assunto(string titulo)
+    {
+        if (string.IsNullOrWhiteSpace(titulo))
+            throw new ArgumentException("Título não pode ser vazio");
+
+        Titulo = titulo;
+    }
+
+    public abstract void ExecutarDemo();
+}
